# server/docs/swagger.yaml
definitions:
  models.Address:
    properties:
      city:
        type: string
      id:
        type: integer
      state:
        type: string
      street:
        type: string
      zipcode:
        type: string
    type: object
  models.Customer:
    properties:
      address:
        $ref: '#/definitions/models.Address'
      address_id:
        type: integer
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phone_number:
        type: string
    type: object
  models.Location:
    properties:
      address:
        $ref: '#/definitions/models.Address'
      address_id:
        type: integer
      id:
        type: integer
      name:
        type: string
    type: object
  models.Rental:
    properties:
      customer:
        $ref: '#/definitions/models.Customer'
      customer_id:
        type: integer
      end_time:
        type: string
      excess_fee:
        type: number
      id:
        type: integer
      rental_fee:
        type: number
      start_time:
        type: string
      trailer:
        $ref: '#/definitions/models.Trailer'
      trailer_id:
        type: integer
    type: object
  models.Trailer:
    properties:
      availability_status:
        type: string
      id:
        type: integer
      location:
        $ref: '#/definitions/models.Location'
      location_id:
        type: integer
      number:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/rental:
    post:
      consumes:
      - application/json
      description: Create a new rental entry in the database
      parameters:
      - description: trailerId
        in: query
        name: trailer_id
        required: true
        type: string
      - description: userId
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Rental'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Create a new rental
      tags:
      - rental
  /api/trailer:
    get:
      description: Fetches a list of all trailers from the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Trailer'
            type: array
      summary: Get all trailers
      tags:
      - trailers
  /api/trailer/{zip}:
    get:
      description: Fetches a trailer based on the zipcode from the database
      parameters:
      - description: Zip
        in: path
        name: zip
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Trailer'
      summary: Get a trailer
      tags:
      - trailer
  /api/user:
    get:
      description: Retrieves all users from the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Customer'
            type: array
        "500":
          description: error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Creates a new user entry in the database
      parameters:
      - description: User Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.Customer'
      produces:
      - application/json
      responses:
        "201":
          description: user_id
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: error
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Creates a new user
      tags:
      - users
  /api/user/{id}:
    get:
      description: Retrieves a user from the database based on the user ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Customer'
        "500":
          description: error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a user by ID
      tags:
      - users
swagger: "2.0"